# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "twitterDemo Api",
    "description": "twitterDemo 相關 Api"
  },
  "host": "fast-tundra-64548.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions":{
    "Bearer":{
      "type": "apiKey",
      "in": "headers",
      "name": "Authorization",
      "description": "JWT Token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/post/{id}/likes": {
      "get": {
        "tags": ["Post"],
        "summary": "取得單則貼文按讚的人",
        "description": "取得單則貼文按讚的人",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [ {
                    "_id": "627100547ef69b72689e67f6",
                    "name": "Michael Jackson",
                    "photo": "https://i.imgur.com/eInPDId.png"
                }]
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401 (未帶 Token)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundToken"
            }
          },
          "401 (Token 失效)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notCorrectToken"
            }
          },
          "401 (Token 未授權)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/invalidToken"
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請帶入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "查無此貼文，請輸入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Post"],
        "summary": "按讚/收回按讚",
        "description": "按讚/收回按讚",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "_id": "62763cb70b6e444e6fd4fbbb",
                    "user": {
                      "_id": "627100547ef69b72689e67f6",
                      "name": "Michael Jackson",
                      "photo": "https://i.imgur.com/eInPDId.png"
                    },
                    "content": "更新內容",
                    "images": [],
                    "likes": [
                      "627100547ef69b72689e67f6"
                    ],
                    "createdAt": "2022-05-07T09:32:39.272Z",
                    "comments": []
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401 (未帶 Token)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundToken"
            }
          },
          "401 (Token 失效)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notCorrectToken"
            }
          },
          "401 (Token 未授權)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/invalidToken"
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請帶入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "查無此貼文，請輸入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/post/{id}/comments": {
      "get": {
        "tags": ["Post"],
        "summary": "取得單則貼文的留言",
        "description": "取得單則貼文的留言",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "example": [{
                    "_id": "62ac4f1b43d95c3ab715e163",
                    "content": "留言內容123",
                    "commenter": {
                        "_id": "62a9c4c2cabb5fb2731803ff",
                        "name": "Michael Jackson",
                        "isThirdPartyLogin": false,
                        "photo": "https://i.imgur.com/eInPDId.png"
                    },
                    "post": "62aae97e072d31d94077a9e6",
                    "createdAt": "2022-06-17T09:53:31.616Z"
                  },
                  {
                    "_id": "62ac442b666939c7ff239902",
                    "content": "留言內容",
                    "commenter": {
                        "_id": "62a9c4c2cabb5fb2731803ff",
                        "name": "Michael Jackson",
                        "isThirdPartyLogin": false,
                        "photo": "https://i.imgur.com/eInPDId.png"
                    },
                    "post": "62aae97e072d31d94077a9e6",
                    "createdAt": "2022-06-17T09:06:51.992Z"
                  }]
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401 (未帶 Token)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundToken"
            }
          },
          "401 (Token 失效)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notCorrectToken"
            }
          },
          "401 (Token 未授權)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/invalidToken"
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請帶入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "查無此貼文，請輸入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Post"],
        "summary": "新增留言",
        "description": "新增留言",
        "security":[{
          "Bearer":[]
        }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "貼文ID"
          },{
            "name": "obj",
            "in": "body",
            "description": "資料格式",
            "required":true,
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "string",
                  "example": "留言內容"
                }
              },
              "required": [
                "content"
              ]
            },
            "type": "string"
            }
        ],
        "responses": {
          "200": {
            "description": "成功回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "example": {
                      "content": "留言內容888",
                      "commenter": {
                          "_id": "62a9c4c2cabb5fb2731803ff",
                          "name": "wendy Chou",
                          "isThirdPartyLogin": false,
                          "photo": "https://i.imgur.com/eInPDId.png"
                      },
                      "post": "62aae97e072d31d94077a9e6",
                      "_id": "62b03bf9c2a3d4e69e426e5a",
                      "createdAt": "2022-06-20T09:20:57.961Z"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (沒有留言內容)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請輸入留言內容"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401 (未帶 Token)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notFoundToken"
            }
          },
          "401 (Token 失效)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/notCorrectToken"
            }
          },
          "401 (Token 未授權)": {
            "description": "失敗回應",
            "schema": {
              "$ref":"#/definitions/invalidToken"
            }
          },
          "400 (ID 格式錯誤)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "請帶入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400 (查無此 ID)": {
            "description": "失敗回應",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "查無此貼文，請輸入正確ID"
                }
              
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "notFoundToken":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "未帶入token"
        }
      
      },
      "xml": {
        "name": "main"
      }
    },
    "notCorrectToken":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "格式錯誤"
        }
      
      }
    },
    "invalidToken":{
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": false
        },
        "message": {
          "type": "string",
          "example": "未授權"
        }
      
      },
      "xml": {
        "name": "main"
      }
    },
...
  }
}